# api-spec.yaml
openapi: 3.0.3
info:
  title: Next Loyalty Program API
  description: 'Unified Loyalty Program API for Next plc omnichannel retail.

    Supports customer registration, points earning/redemption, tier management,

    notifications, staff operations, and admin configuration.

    '
  version: 1.0.0
  contact:
    name: Next Loyalty Engineering
    email: loyalty-engineering@next.co.uk
servers:
- url: https://loyalty.next.co.uk/api/v1
  description: Production
- url: https://staging.loyalty.next.co.uk/api/v1
  description: Staging
- url: http://localhost:3000/api/v1
  description: Local Development
security:
- bearerAuth: []
...
paths:
...
  /tiers:
    get:
      operationId: listTiers
      summary: List all loyalty tiers
      description: Retrieve all configured loyalty tiers with thresholds, multipliers,
        and benefits.
      tags:
      - Tiers
      security: []
      responses:
        '200':
          description: Tier list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Tier'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /staff/customers/search:
    get:
      operationId: staffSearchCustomers
      summary: Search customers (staff)
      description: Search loyalty customers by phone, email, name, or loyalty ID.
        Requires staff role. PII masked in results.
      tags:
      - Staff
      security:
      - bearerAuth: []
      parameters:
      - name: q
        in: query
        required: true
        schema:
          type: string
          minLength: 2
        description: Search query value
      - name: field
        in: query
        required: true
        schema:
          type: string
          enum:
          - phone
          - email
          - name
          - loyalty_id
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        customerId:
                          type: string
                          format: uuid
                        name:
                          type: string
                        maskedEmail:
                          type: string
                          example: j***@example.com
                        maskedPhone:
                          type: string
                          example: +44***0123
                        tier:
                          type: string
                        pointsBalance:
                          type: integer
                        memberSince:
                          type: string
                          format: date
                  totalResults:
                    type: integer
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /staff/customers/{id}/loyalty-summary:
    get:
      operationId: getCustomerLoyaltySummary
      summary: Get customer loyalty summary (staff)
      description: Full loyalty summary for POS display. Includes tier, balance, recent
        transactions, and personalisation hints.
      tags:
      - Staff
      security:
      - bearerAuth: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Loyalty summary
          content:
            application/json:
              schema:
                type: object
                properties:
                  customerId:
                    type: string
                    format: uuid
                  name:
                    type: string
                  tier:
                    type: string
                  tierBadgeColor:
                    type: string
                    example: '#FFD700'
                  pointsBalance:
                    type: integer
                  monetaryEquivalent:
                    type: number
                    format: double
                  lifetimePoints:
                    type: integer
                  memberSince:
                    type: string
                    format: date
                  recentTransactions:
                    type: array
                    maxItems: 3
                    items:
                      $ref: '#/components/schemas/Transaction'
                  personalisationHints:
                    type: array
                    items:
                      type: string
                    example:
                    - Gold member — eligible for 10% bonus
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /staff/adjustments:
    post:
      operationId: createPointsAdjustment
      summary: Create manual points adjustment
      description: Add or deduct points manually. Adjustments >500 points require
        manager approval. All adjustments are audit-logged.
      tags:
      - Staff
      security:
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - customerId
              - action
              - points
              - reason
              properties:
                customerId:
                  type: string
                  format: uuid
                action:
                  type: string
                  enum:
                  - add
                  - deduct
                points:
                  type: integer
                  minimum: 1
                  example: 200
                reason:
                  type: string
                  enum:
                  - system_error_correction
                  - goodwill_gesture
                  - promotional_credit
                  - fraud_correction
                  - other
                notes:
                  type: string
                  maxLength: 500
                  description: Required if reason is 'other'
      responses:
        '201':
          description: Adjustment created
          content:
            application/json:
              schema:
                type: object
                properties:
                  adjustmentId:
                    type: string
                    format: uuid
                  status:
                    type: string
                    enum:
                    - approved
                    - pending_approval
                  pointsAdjusted:
                    type: integer
                  newBalance:
                    type: integer
                    nullable: true
                    description: Null if pending approval
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /staff/adjustments/{id}/approve:
    patch:
      operationId: approvePointsAdjustment
      summary: Approve points adjustment
      description: Approve a pending points adjustment. Requires loyalty_manager role.
      tags:
      - Staff
      security:
      - bearerAuth: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - approved
              properties:
                approved:
                  type: boolean
                reviewNotes:
                  type: string
                  maxLength: 500
      responses:
        '200':
          description: Adjustment reviewed
          content:
            application/json:
              schema:
                type: object
                properties:
                  adjustmentId:
                    type: string
                    format: uuid
                  status:
                    type: string
                    enum:
                    - approved
                    - rejected
                  newBalance:
                    type: integer
                    nullable: true
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /webhooks/ecommerce:
    post:
      operationId: receiveEcommerceWebhook
      summary: Receive e-commerce event
      description: Webhook receiver for e-commerce platform purchase/refund events.
        Validates HMAC-SHA256 signature. Idempotent by order_id + event_type.
      tags:
      - Webhooks
      security:
      - apiKeyAuth: []
      parameters:
      - name: X-Webhook-Signature
        in: header
        required: true
        schema:
          type: string
        description: HMAC-SHA256 signature of request body
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - eventType
              - orderId
              - customerEmail
              - totalAmount
              properties:
                eventType:
                  type: string
                  enum:
                  - order.completed
                  - order.refunded
                orderId:
                  type: string
                  example: ECOM-2026-789012
                customerEmail:
                  type: string
                  format: email
                totalAmount:
                  type: number
                  format: double
                  example: 149.99
                currency:
                  type: string
                  example: GBP
                items:
                  type: array
                  items:
                    type: object
                    properties:
                      sku:
                        type: string
                      name:
                        type: string
                      category:
                        type: string
                      amount:
                        type: number
                        format: double
                timestamp:
                  type: string
                  format: date-time
      responses:
        '202':
          description: Webhook accepted for processing
          content:
            application/json:
              schema:
                type: object
                properties:
                  accepted:
                    type: boolean
                    example: true
                  correlationId:
                    type: string
                    format: uuid
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /webhooks/pos:
    post:
      operationId: receivePosWebhook
      summary: Receive POS transaction event
      description: Webhook receiver for POS system purchase/refund events. Validates
        API key. Idempotent by transaction_id.
      tags:
      - Webhooks
      security:
      - apiKeyAuth: []
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - eventType
              - transactionId
              - customerId
              - totalAmount
              - storeId
              properties:
                eventType:
                  type: string
                  enum:
                  - transaction.completed
                  - transaction.refunded
                transactionId:
                  type: string
                  example: POS-STORE42-20260518-001
                customerId:
                  type: string
                  format: uuid
                totalAmount:
                  type: number
                  format: double
                  example: 85.0
                storeId:
                  type: string
                  example: STORE-042
                terminalId:
                  type: string
                  example: POS-STORE42-01
                items:
                  type: array
                  items:
                    type: object
                    properties:
                      sku:
                        type: string
                      category:
                        type: string
                      amount:
                        type: number
                        format: double
                timestamp:
                  type: string
                  format: date-time
      responses:
        '202':
          description: Event accepted for processing
          content:
            application/json:
              schema:
                type: object
                properties:
                  accepted:
                    type: boolean
                    example: true
                  correlationId:
                    type: string
                    format: uuid
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token (15-minute expiry)
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for POS terminals and e-commerce integrations
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        format: uuid
      description: Unique key to ensure idempotent processing of write operations
...
  schemas:
...
    Transaction:
      type: object
      properties:
        transactionId:
          type: string
          format: uuid
        type:
          type: string
          enum:
          - earn
          - redeem
          - adjustment
          - bonus
          - expired
          - reversal
        points:
          type: integer
          description: Positive for credits, negative for debits
          example: 76
        runningBalance:
          type: integer
          example: 2576
        referenceId:
          type: string
          description: Order ID, adjustment ID, or campaign ID
          example: ORD-2026-001234
        channel:
          type: string
          enum:
          - online
          - in_store
          - app
          - system
        description:
          type: string
          example: Purchase at Next Oxford Street
        createdAt:
          type: string
          format: date-time
      example:
        transactionId: f1e2d3c4-b5a6-7890-fedc-ba0987654321
        type: earn
        points: 76
        runningBalance: 2576
        referenceId: ORD-2026-001234
        channel: online
        description: Online purchase
        createdAt: '2026-05-18T14:30:00Z'
    Tier:
      type: object
      properties:
        tierId:
          type: string
          format: uuid
        name:
          type: string
          example: Gold
        threshold:
          type: integer
          description: Lifetime points required to reach this tier
          example: 5000
        earnRateMultiplier:
          type: number
          format: double
          example: 1.5
        badgeColor:
          type: string
          example: '#FFD700'
        benefits:
          type: array
          items:
            type: string
          example:
          - 1.5x points on all purchases
          - Early access to sales (24 hours)
          - Free delivery on all orders
      example:
        tierId: c3d4e5f6-a7b8-9012-cdef-345678901234
        name: Gold
        threshold: 5000
        earnRateMultiplier: 1.5
        badgeColor: '#FFD700'
        benefits:
        - 1.5x points on all purchases
        - Early access to sales (24 hours)
        - Free delivery on all orders
...
    ProblemDetails:
      type: object
      description: RFC 7807 Problem Details for HTTP APIs
      properties:
        type:
          type: string
          format: uri
          description: URI reference identifying the problem type
          example: https://loyalty.next.co.uk/problems/validation-error
        title:
          type: string
          description: Short human-readable summary
          example: Validation Error
        status:
          type: integer
          description: HTTP status code
          example: 422
        detail:
          type: string
          description: Human-readable explanation specific to this occurrence
          example: The 'email' field must be a valid email address
        instance:
          type: string
          format: uri
          description: URI reference identifying the specific occurrence
          example: /api/v1/auth/register
        traceId:
          type: string
          description: Correlation ID for distributed tracing
          example: abc123def456
        errors:
          type: array
          description: Detailed field-level validation errors
          items:
            type: object
            properties:
              field:
                type: string
                example: email
              message:
                type: string
                example: Must be a valid email address
              code:
                type: string
                example: INVALID_FORMAT
      example:
        type: https://loyalty.next.co.uk/problems/validation-error
        title: Validation Error
        status: 422
        detail: One or more fields failed validation
        instance: /api/v1/auth/register
        traceId: abc123def456
        errors:
        - field: email
          message: Must be a valid email address
          code: INVALID_FORMAT
...
  responses:
    BadRequest:
      description: Bad Request — malformed request syntax or invalid parameters
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            type: https://loyalty.next.co.uk/problems/bad-request
            title: Bad Request
            status: 400
            detail: Request body is malformed or missing required fields
            traceId: trace-001
    Unauthorized:
      description: Unauthorized — missing or invalid authentication credentials
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            type: https://loyalty.next.co.uk/problems/unauthorized
            title: Unauthorized
            status: 401
            detail: Access token is missing, expired, or invalid
            traceId: trace-002
    Forbidden:
      description: Forbidden — authenticated but insufficient permissions
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            type: https://loyalty.next.co.uk/problems/forbidden
            title: Forbidden
            status: 403
            detail: You do not have permission to perform this action
            traceId: trace-003
    NotFound:
      description: Not Found — requested resource does not exist
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            type: https://loyalty.next.co.uk/problems/not-found
            title: Not Found
            status: 404
            detail: The requested resource was not found
            traceId: trace-004
    UnprocessableEntity:
      description: Unprocessable Entity — request understood but validation failed
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            type: https://loyalty.next.co.uk/problems/validation-error
            title: Validation Error
            status: 422
            detail: One or more fields failed validation
            traceId: trace-005
            errors:
            - field: email
              message: Email already registered
              code: DUPLICATE
    InternalServerError:
      description: Internal Server Error — unexpected server-side failure
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            type: https://loyalty.next.co.uk/problems/internal-error
            title: Internal Server Error
            status: 500
            detail: An unexpected error occurred. Please try again later.
            traceId: trace-006